@@ -13,40 +13,40 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "7.2.12",
-    "@angular/common": "7.2.12",
-    "@angular/compiler": "7.2.12",
-    "@angular/core": "7.2.12",
-    "@angular/forms": "7.2.12",
-    "@angular/http": "7.2.12",
-    "@angular/platform-browser": "7.2.12",
-    "@angular/platform-browser-dynamic": "7.2.12",
-    "@angular/router": "7.2.12",
+    "@angular/animations": "8.1.3",
+    "@angular/common": "8.1.3",
+    "@angular/compiler": "8.1.3",
+    "@angular/core": "8.1.3",
+    "@angular/forms": "8.1.3",
+    "@angular/platform-browser": "8.1.3",
+    "@angular/platform-browser-dynamic": "8.1.3",
+    "@angular/router": "8.1.3",
     "core-js": "2.6.0",
-    "ngx-quill": "3.3.0",
-    "rxjs": "6.4.0",
-    "zone.js": "0.9.0"
+    "ngx-quill": "7.1.2",
+    "quill": "^1.3.6",
+    "rxjs": "6.5.2",
+    "zone.js": "0.9.1"
   },
   "devDependencies": {
-    "@angular/cli": "7.3.8",
-    "@angular/compiler-cli": "7.2.12",
-    "@angular/language-service": "7.2.12",
-    "@types/jasmine": "3.3.12",
+    "@angular/cli": "8.1.3",
+    "@angular/compiler-cli": "8.1.3",
+    "@angular/language-service": "8.1.3",
+    "@types/jasmine": "3.3.16",
     "@types/jasminewd2": "2.0.6",
-    "@types/node": "11.13.0",
-    "codelyzer": "5.0.0",
+    "@types/node": "12.6.8",
+    "codelyzer": "5.1.0",
     "jasmine-core": "3.4.0",
     "jasmine-spec-reporter": "4.2.1",
-    "karma": "4.0.1",
-    "karma-chrome-launcher": "2.2.0",
+    "karma": "4.2.0",
+    "karma-chrome-launcher": "3.0.0",
     "karma-cli": "2.0.0",
-    "karma-coverage-istanbul-reporter": "2.0.5",
+    "karma-coverage-istanbul-reporter": "2.1.0",
     "karma-jasmine": "2.0.1",
     "karma-jasmine-html-reporter": "1.4.0",
     "protractor": "5.4.2",
-    "ts-node": "8.0.3",
-    "tslint": "5.15.0",
-    "typescript": "3.1.2",
-    "@angular-devkit/build-angular": "0.13.8"
+    "ts-node": "8.3.0",
+    "tslint": "5.18.0",
+    "typescript": "3.4.5",
+    "@angular-devkit/build-angular": "0.801.3"
   }
 }
